\name{BondsData}

\alias{BondsData}


\alias{bundesbankNSS}
\alias{mk.zero2}
\alias{mk.maturity}    
            

\title{Bonds Data Sets}


\description{

    A collection and description of data sets 
    from the Rmetrics Package fBonds.
    \cr
    
    The data sets are:
    
    \tabular{ll}{   
    \code{bundesbankNSS} \tab Nelson-Siegel-Svensson Coefficients. }
    
}


\details{
            
    \bold{Bundesbank Nelson-Siegel-Svensson Coefficients:}
    \cr\cr
    Coefficients for the Nelson-Siegel-Svensson yield curve
    from the German Bundesbank.
    \cr
    
    \bold{:}
    \cr\cr
    \code{mk.zero2} is a data set with a 67 x 55 values 
    representing the zero-coupon yield curve from August 1985 
    to February 1991.\cr
    \code{mk.maturity} a numeric vector of length 55, giving 
    the fifty-five maturities in terms of years for the term 
    structure. 

}


\references{

McCulloch J. H. (1990);
    \emph{US term structure data: 1946-87}, 
    Handbook of Monetary Economics, Friedman B.M. and Hahn F.H. (eds.), 
    Elsevier Science.  
  
McCulloch J. H. and Kwon, H.C. (1993);
    \emph{US term structure data: 1947-1991}, 
    Working Paper No. 93-6, Department of Economics, 
    Ohio State University.
      
}


\keyword{datasets}

