\name{RforQuantsData}

\alias{RforQuantsData}

\alias{bmwRet}



\title{Time Series Data Sets fro the ebook RforQuants} 


\description{

    Data sets used in examples of several Rmetrics packages.
    \cr
    
    The data sets are:
    
    \tabular{ll}{   
    \code{PensionFund} \tab Pension Fund Portfolio.}
   
}
    
    
\details{

       
    \bold{Pension Fund Benchmark:}\cr
    \cr
    \code{PensionFund}
    iswhat, explain it here.\cr
    The data set ranges from 20xx-01-01 to 20yy-12-31.\cr
    The columns are named: XXX.

}


\format{
    
    All files are in CSV Excel spreadsheet format. The
    delimiter is a semicolon. 
    
}


\references{


Pictet (2005);
    \emph{Pension Fund Benchmark},
    www.pictet.ch . 

}


\examples{
## Load Example Data Set:
   data(PensionFund)
   head(PensionFund)
}
 
 
\keyword{datasets}

