\name{undocumentedTimeSeriesData}

\alias{undocumentedTimeSeriesData}

\alias{DowJones}
\alias{GDP}
\alias{GNP}
\alias{HedgeFund}
\alias{.HedgeFund1}
\alias{.HedgeFund2}
\alias{IDUdata}
\alias{Capitalization}
\alias{PensionFund}
\alias{swissEconomy}
\alias{air}
\alias{cars2}
\alias{test1}
\alias{test2}

\title{Undocumented Time Series Data Sets}
\description{These data sets need documentation.}
